Codahale -> Dropwizard updates.
When updating Guava to a later version, we discovered that there were
now conflicting classes for various Codahale metrics libraries.  This
led us to discover that the project started out as one of Coda's
personal projects; it used class namespace `com.codahale.metrics`.  He
was working at Yammer at the time, and the project became sufficiently
important for them to change its packaging to be associated with Yammer.
When they did this, they _changed_ the class namespace to
`com.yammer.metrics`.  Later, the project became sufficiently valuable
to the community that they open-sourced it.  On doing this, though, they
*changed the class namespace _back_* to `com.codahale.metrics` despite
the fact that the artifact `groupId` changed yet again to
`io.dropwizard.metrics`.  (wtf!)  We want only the modern Dropwizard
artifacts.

[See more here.][1]

[1]: https://groups.google.com/d/msg/dropwizard-user/1usH7frpnZE/RSQUsOBFMsoJ