jetty json request log: refactor to use otl definitions; rework factory
The old interface is not flexible enough.  You could only create objects that
extend RequestLogEvent, making it hard to add new fields, and harder yet to
compose with other OTL.

Now, you just replace the JsonRequestLog with a subclass entirely, and ditch the
factory.

I verified that Frontdoor is the only project that ever used this feature.